+++ cmd/golivepilot/main.go
@@ -22,7 +22,7 @@ import (
 )
 
 var (
-	version = "0.1.46"
+	version = "0.1.47"
 )
 
 func main() {